Move showErrorDetailsDialog() to base
