+++ app/src/main/java/io/github/wulkanowy/ui/base/BaseDialogFragment.kt
@@ -1,0 +1,27 @@
+package io.github.wulkanowy.ui.base
+
+import android.widget.Toast
+import dagger.android.support.DaggerAppCompatDialogFragment
+
+abstract class BaseDialogFragment : DaggerAppCompatDialogFragment(), BaseView {
+
+    override fun showError(text: String, error: Throwable) {
+        showMessage(text)
+    }
+
+    override fun showMessage(text: String) {
+        Toast.makeText(context, text, Toast.LENGTH_LONG).show()
+    }
+
+    override fun showExpiredDialog() {
+        (activity as? BaseActivity<*>)?.showExpiredDialog()
+    }
+
+    override fun openClearLoginView() {
+        (activity as? BaseActivity<*>)?.openClearLoginView()
+    }
+
+    override fun showErrorDetailsDialog(error: Throwable) {
+        ErrorDialog.newInstance(error).show(childFragmentManager, error.toString())
+    }
+}
