+++ app/src/main/java/io/github/wulkanowy/ui/base/BaseFragment.kt
@@ -13,14 +13,14 @@ abstract class BaseFragment : DaggerFragment(), BaseView {
     override fun showError(text: String, error: Throwable) {
         if (messageContainer != null) {
             Snackbar.make(messageContainer!!, text, LENGTH_LONG)
-                .setAction(R.string.all_details) { if (isAdded) showErrorDialog(error) }
+                .setAction(R.string.all_details) { if (isAdded) showErrorDetailsDialog(error) }
                 .show()
         } else {
             (activity as? BaseActivity<*>)?.showError(text, error)
         }
     }
 
-    fun showErrorDialog(error: Throwable) {
+    override fun showErrorDetailsDialog(error: Throwable) {
         ErrorDialog.newInstance(error).show(childFragmentManager, error.toString())
     }
