+++ app/src/main/java/io/github/wulkanowy/ui/base/BaseView.kt
@@ -9,4 +9,6 @@ interface BaseView {
     fun showExpiredDialog()
 
     fun openClearLoginView()
+
+    fun showErrorDetailsDialog(error: Throwable)
 }
