+++ app/src/main/java/io/github/wulkanowy/ui/modules/attendance/AttendanceView.kt
@@ -28,8 +28,6 @@ interface AttendanceView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
