+++ app/src/main/java/io/github/wulkanowy/ui/modules/exam/ExamView.kt
@@ -25,8 +25,6 @@ interface ExamView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
