+++ app/src/main/java/io/github/wulkanowy/ui/modules/grade/statistics/GradeStatisticsView.kt
@@ -36,8 +36,6 @@ interface GradeStatisticsView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
