+++ app/src/main/java/io/github/wulkanowy/ui/modules/homework/HomeworkView.kt
@@ -23,8 +23,6 @@ interface HomeworkView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
