+++ app/src/main/java/io/github/wulkanowy/ui/modules/luckynumber/LuckyNumberView.kt
@@ -19,8 +19,6 @@ interface LuckyNumberView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
