+++ app/src/main/java/io/github/wulkanowy/ui/modules/message/preview/MessagePreviewView.kt
@@ -33,8 +33,6 @@ interface MessagePreviewView : BaseView {
 
     fun setErrorRetryCallback(callback: () -> Unit)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showOptions(show: Boolean)
 
     fun setDeletedOptionsLabels()
