+++ app/src/main/java/io/github/wulkanowy/ui/modules/message/tab/MessageTabFragment.kt
@@ -112,10 +112,6 @@ class MessageTabFragment : BaseFragment(), MessageTabView {
         messageTabErrorMessage.text = message
     }
 
-    override fun showErrorDetailsDialog(error: Throwable) {
-        showErrorDialog(error)
-    }
-
     override fun showRefresh(show: Boolean) {
         messageTabSwipe.isRefreshing = show
     }
