+++ app/src/main/java/io/github/wulkanowy/ui/modules/note/NoteView.kt
@@ -22,8 +22,6 @@ interface NoteView : BaseView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
