+++ app/src/main/java/io/github/wulkanowy/ui/modules/schoolandteachers/school/SchoolView.kt
@@ -18,8 +18,6 @@ interface SchoolView : BaseView, SchoolAndTeachersChildView {
 
     fun setErrorDetails(message: String)
 
-    fun showErrorDetailsDialog(error: Throwable)
-
     fun showProgress(show: Boolean)
 
     fun enableSwipe(enable: Boolean)
