+++ app/src/main/java/io/github/wulkanowy/ui/modules/schoolandteachers/teacher/TeacherFragment.kt
@@ -86,10 +86,6 @@ class TeacherFragment : BaseFragment(), TeacherView, MainView.TitledView,
         teacherErrorMessage.text = message
     }
 
-    override fun showErrorDetailsDialog(error: Throwable) {
-        showErrorDialog(error)
-    }
-
     override fun showProgress(show: Boolean) {
         teacherProgress.visibility = if (show) VISIBLE else GONE
     }
